fix: 修复编译错误
- 修复 MainActivity setContent 导入问题
- 修复 VideoItem @Composable 重复注解
- 修复 LinearProgressIndicator progress lambda 参数
- 修复 HorizontalDivider -> Divider
- 修复 selectedDirectoryUri -> selectedFolderUri

+++ app/src/main/java/com/dva/app/presentation/settings/ModelsScreen.kt
@@ -200,7 +200,7 @@ fun ModelCard(
             if (model.isDownloading) {
                 Spacer(modifier = Modifier.height(12.dp))
                 LinearProgressIndicator(
-                    progress = { model.downloadProgress / 100f },
+                    progress = { model.downloadProgress.toFloat() / 100f },
                     modifier = Modifier.fillMaxWidth(),
                 )
                 Spacer(modifier = Modifier.height(4.dp))
